+++ bundle.ts
@@ -1,4 +1,3 @@
-import { buildTeloMisikeke } from "./telo-misikeke/build.ts";
 import { buildDictionary } from "./dictionary/build.ts";
 import { debounce } from "./src/misc.ts";
 import { build } from "esbuild";
@@ -14,6 +13,8 @@ const WATCH = [
   "./project-data.json",
 ];
 async function buildIloToken(minify: boolean): Promise<void> {
+  console.log("Building dictionary...");
+  await buildDictionary();
   console.log("Building main.js...");
   await build({
     entryPoints: ["./src/main.ts"],
@@ -29,8 +30,6 @@ async function buildIloToken(minify: boolean): Promise<void> {
 if (import.meta.main) {
   switch (Deno.args[0]) {
     case "build": {
-      console.log("Building dictionary and telo misikeke...");
-      await Promise.all([buildDictionary(), buildTeloMisikeke()]);
       await buildIloToken(true);
       break;
     }
@@ -38,8 +37,6 @@ if (import.meta.main) {
       console.log("Press ctrl+c to exit.");
       const builder = debounce(async () => {
         try {
-          console.log("Building dictionary...");
-          await buildDictionary();
           await buildIloToken(false);
         } catch (error) {
           console.error(error);