+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
 <!--
-NOTE:
+NOTE: Before publishing:
 - update dependencies before publishing
+- run `deno task update` to update telo misikeke and Linku data
 - set parameters on `project-data.json`
 -->